Modify : VAC 패턴 적용 및 타입 재지정
