+++ src/api/select.ts
@@ -1,8 +1,9 @@
+import { ISelectedInfo } from '@src/types/select';
 import API from './api';
 import client from './client';
 
 const SelectAPI = {
-  postSelected: (data: ISelect) => {
+  postSelected: (data: ISelectedInfo) => {
     return client.post(`${API.postSelected}`, data);
   },
 };
