+++ src/components/result/Clipboard.tsx
@@ -1,55 +1,21 @@
-import styled from '@emotion/styled';
+import { IClipboardProps } from '@src/types/result';
 import React, { useState } from 'react';
+import ClipboardView from './Views/ClipboardView';
 
 const Clipboard = () => {
   const [copyUrl, setCopyUrl] = useState<boolean>(false);
 
-  const SERVICE_URL = 'https://mytamla.netlify.app';
+  const ClipboardProps: IClipboardProps = {
+    copyUrl,
+    serviceUrl: 'https://mytamla.netlify.app',
+    handleCopyClipBoard: () => {
+      setCopyUrl(true);
 
-  const handleCopyClipBoard = () => {
-    setCopyUrl(true);
-
-    return navigator.clipboard.writeText(SERVICE_URL);
+      return navigator.clipboard.writeText('https://mytamla.netlify.app');
+    },
   };
 
-  return (
-    <StClipboard>
-      <StUrl>{SERVICE_URL}</StUrl>
-      <StCopyBtn isCopied={copyUrl} onClick={handleCopyClipBoard}>
-        복사
-      </StCopyBtn>
-    </StClipboard>
-  );
+  return <ClipboardView {...ClipboardProps} />;
 };
-
-const StClipboard = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  background: ${({ theme }) => theme.color.gray400};
-  border: 1px solid black;
-  border-radius: 15px;
-`;
-
-const StUrl = styled.div`
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  padding: 20px;
-  width: 80%;
-  height: 55px;
-`;
-
-const StCopyBtn = styled.button<{ isCopied: boolean }>`
-  width: 20%;
-  height: 55px;
-  color: ${({ theme, isCopied }) => (isCopied ? theme.color.backgroundColor : theme.color.whiteColor)};
-  background: ${({ theme, isCopied }) => (isCopied ? theme.color.mainColor : theme.color.gray500)};
-  border: none;
-  border-top-right-radius: 15px;
-  border-bottom-right-radius: 15px;
-  font-size: 16px;
-`;
 
 export default Clipboard;
