+++ src/utils/converToArray.ts
@@ -1,3 +1,5 @@
-export const convertToArray = (object: ISelect) => {
+import { ISelectedInfo } from '@src/types/select';
+
+export const convertToArray = (object: ISelectedInfo) => {
   return Object.values(object);
 };
